# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sanity API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "PersonalBestSortInput": {
        "type": "string",
        "enum": [
          "content",
          "scale",
          "submittedAt",
          "time"
        ]
      },
      "MemberSortInput": {
        "type": "string",
        "enum": [
          "clanPoints",
          "diaryPoints",
          "displayName",
          "joinedAt",
          "masterDiaries",
          "rank",
          "totalEhb",
          "totalEhp",
          "totalExp",
          "totalLevel"
        ]
      },
...
      "Page": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "offset": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        },
        "required": [
          "limit",
          "offset",
          "total",
          "hasNext",
          "hasPrevious"
        ],
        "additionalProperties": false
      },
      "MemberRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ],
        "additionalProperties": false
      },
...
      "RankedPersonalBest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "position": {
            "type": "number"
          },
          "content": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "imageUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "imageUrl"
            ],
            "additionalProperties": false
          },
          "scale": {
            "type": "number"
          },
          "timeSeconds": {
            "type": "number"
          },
          "imageUrl": {
            "nullable": true,
            "type": "string"
          },
          "submittedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "submittedBy": {
            "$ref": "#/components/schemas/MemberRef"
          },
          "team": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRef"
            }
          }
        },
        "required": [
          "id",
          "position",
          "content",
          "scale",
          "timeSeconds",
          "imageUrl",
          "submittedAt",
          "submittedBy",
          "team"
        ],
        "additionalProperties": false
      },
      "MemberDiary": {
        "type": "object",
        "properties": {
          "content": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "imageUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "imageUrl"
            ],
            "additionalProperties": false
          },
          "scale": {
            "type": "number"
          },
          "pb": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "timeSeconds": {
                "type": "number"
              },
              "imageUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "timeSeconds",
              "imageUrl"
            ],
            "additionalProperties": false
          },
          "tier": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "points": {
                "type": "number"
              }
            },
            "required": [
              "id",
              "name",
              "points"
            ],
            "additionalProperties": false
          },
          "nextTier": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "points": {
                "type": "number"
              },
              "timeSeconds": {
                "type": "number"
              }
            },
            "required": [
              "id",
              "name",
              "points",
              "timeSeconds"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "content",
          "scale",
          "pb",
          "tier",
          "nextTier"
        ],
        "additionalProperties": false
      },
...
      "RankSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "iconUrl"
        ],
        "additionalProperties": false
      },
      "Computed": {
        "type": "object",
        "properties": {
          "metric": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "rank": {
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "metric",
          "value",
          "rank"
        ],
        "additionalProperties": false
      },
      "Skill": {
        "type": "object",
        "properties": {
          "skill": {
            "type": "string"
          },
          "experience": {
            "type": "number"
          },
          "level": {
            "type": "number"
          },
          "rank": {
            "nullable": true,
            "type": "number"
          },
          "ehp": {
            "type": "number"
          }
        },
        "required": [
          "skill",
          "experience",
          "level",
          "rank",
          "ehp"
        ],
        "additionalProperties": false
      },
      "Boss": {
        "type": "object",
        "properties": {
          "boss": {
            "type": "string"
          },
          "kills": {
            "type": "number"
          },
          "rank": {
            "nullable": true,
            "type": "number"
          },
          "ehb": {
            "type": "number"
          }
        },
        "required": [
          "boss",
          "kills",
          "rank",
          "ehb"
        ],
        "additionalProperties": false
      },
      "Activity": {
        "type": "object",
        "properties": {
          "activity": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "rank": {
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "activity",
          "score",
          "rank"
        ],
        "additionalProperties": false
      },
      "Snapshot": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "computed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Computed"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skill"
            }
          },
          "bosses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Boss"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        },
        "required": [
          "createdAt",
          "computed",
          "skills",
          "bosses",
          "activities"
        ],
        "additionalProperties": false
      },
      "Member": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "displayName": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "discord": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^\\d+$"
              },
              "avatarUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "avatarUrl"
            ],
            "additionalProperties": false
          },
          "rsn": {
            "type": "object",
            "properties": {
              "main": {
                "nullable": true,
                "type": "string"
              },
              "alt": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "main",
              "alt"
            ],
            "additionalProperties": false
          },
          "membership": {
            "type": "object",
            "properties": {
              "active": {
                "type": "boolean"
              },
              "joinedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "leftAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "rank": {
                "$ref": "#/components/schemas/RankSummary"
              },
              "points": {
                "type": "number"
              }
            },
            "required": [
              "active",
              "joinedAt",
              "leftAt",
              "rank",
              "points"
            ],
            "additionalProperties": false
          },
          "diary": {
            "type": "object",
            "properties": {
              "points": {
                "type": "number"
              },
              "masterDiaries": {
                "type": "number"
              },
              "claimedTier": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "points",
              "masterDiaries",
              "claimedTier"
            ],
            "additionalProperties": false
          },
          "wom": {
            "nullable": true,
            "type": "object",
            "properties": {
              "playerId": {
                "type": "number"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "build": {
                "type": "string"
              },
              "status": {
                "nullable": true,
                "type": "string"
              },
              "totalLevel": {
                "nullable": true,
                "type": "number"
              },
              "totalExp": {
                "type": "number"
              },
              "totalEhp": {
                "type": "number"
              },
              "totalEhb": {
                "type": "number"
              },
              "timeToMax": {
                "nullable": true,
                "type": "number"
              },
              "timeTo200m": {
                "nullable": true,
                "type": "number"
              },
              "updatedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "lastChangedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "playerId",
              "username",
              "displayName",
              "type",
              "build",
              "status",
              "totalLevel",
              "totalExp",
              "totalEhp",
              "totalEhb",
              "timeToMax",
              "timeTo200m",
              "updatedAt",
              "lastChangedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "displayName",
          "nationality",
          "discord",
          "rsn",
          "membership",
          "diary",
          "wom"
        ],
        "additionalProperties": false
      },
      "MemberProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "displayName": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "discord": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^\\d+$"
              },
              "avatarUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "avatarUrl"
            ],
            "additionalProperties": false
          },
          "rsn": {
            "type": "object",
            "properties": {
              "main": {
                "nullable": true,
                "type": "string"
              },
              "alt": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "main",
              "alt"
            ],
            "additionalProperties": false
          },
          "membership": {
            "type": "object",
            "properties": {
              "active": {
                "type": "boolean"
              },
              "joinedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "leftAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "rank": {
                "$ref": "#/components/schemas/RankSummary"
              },
              "points": {
                "type": "number"
              }
            },
            "required": [
              "active",
              "joinedAt",
              "leftAt",
              "rank",
              "points"
            ],
            "additionalProperties": false
          },
          "diary": {
            "type": "object",
            "properties": {
              "points": {
                "type": "number"
              },
              "masterDiaries": {
                "type": "number"
              },
              "claimedTier": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "points",
              "masterDiaries",
              "claimedTier"
            ],
            "additionalProperties": false
          },
          "progression": {
            "type": "object",
            "properties": {
              "eligibleRank": {
                "nullable": true,
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RankSummary"
                  }
                ]
              },
              "nextRank": {
                "type": "object",
                "properties": {
                  "points": {
                    "nullable": true,
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "iconUrl": {
                        "nullable": true,
                        "type": "string"
                      },
                      "requirements": {
                        "type": "object",
                        "properties": {
                          "clanPoints": {
                            "type": "number"
                          },
                          "diaryPoints": {
                            "type": "number"
                          },
                          "masterDiaries": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "clanPoints",
                          "diaryPoints",
                          "masterDiaries"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "iconUrl",
                      "requirements"
                    ],
                    "additionalProperties": false
                  },
                  "diary": {
                    "nullable": true,
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "iconUrl": {
                        "nullable": true,
                        "type": "string"
                      },
                      "requirements": {
                        "type": "object",
                        "properties": {
                          "clanPoints": {
                            "type": "number"
                          },
                          "diaryPoints": {
                            "type": "number"
                          },
                          "masterDiaries": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "clanPoints",
                          "diaryPoints",
                          "masterDiaries"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "iconUrl",
                      "requirements"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "points",
                  "diary"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "eligibleRank",
              "nextRank"
            ],
            "additionalProperties": false
          },
          "wom": {
            "nullable": true,
            "type": "object",
            "properties": {
              "playerId": {
                "type": "number"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "build": {
                "type": "string"
              },
              "status": {
                "nullable": true,
                "type": "string"
              },
              "totalLevel": {
                "nullable": true,
                "type": "number"
              },
              "totalExp": {
                "type": "number"
              },
              "totalEhp": {
                "type": "number"
              },
              "totalEhb": {
                "type": "number"
              },
              "timeToMax": {
                "nullable": true,
                "type": "number"
              },
              "timeTo200m": {
                "nullable": true,
                "type": "number"
              },
              "updatedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "lastChangedAt": {
                "nullable": true,
                "type": "string",
                "format": "date-time"
              },
              "latestSnapshot": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Snapshot"
                  }
                ]
              }
            },
            "required": [
              "playerId",
              "username",
              "displayName",
              "type",
              "build",
              "status",
              "totalLevel",
              "totalExp",
              "totalEhp",
              "totalEhb",
              "timeToMax",
              "timeTo200m",
              "updatedAt",
              "lastChangedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "displayName",
          "nationality",
          "discord",
          "rsn",
          "membership",
          "diary",
          "progression",
          "wom"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/members": {
      "get": {
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "active",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 50
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "default": "clanPoints",
              "allOf": [
                {
                  "$ref": "#/components/schemas/MemberSortInput"
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "in": "query",
            "name": "order",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Member"
                      }
                    },
                    "page": {
                      "$ref": "#/components/schemas/Page"
                    }
                  },
                  "required": [
                    "items",
                    "page"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/members/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfile"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/members/{id}/diary": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDiary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/members/{id}/personal-bests": {
      "get": {
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10
            },
            "in": "query",
            "name": "top",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "contentId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "diary",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "scale",
            "required": false
          },
          {
            "schema": {
              "default": "submittedAt",
              "allOf": [
                {
                  "$ref": "#/components/schemas/PersonalBestSortInput"
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "in": "query",
            "name": "order",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RankedPersonalBest"
                      }
                    },
                    "page": {
                      "$ref": "#/components/schemas/Page"
                    }
                  },
                  "required": [
                    "items",
                    "page"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}